+++ pages/blog/viewAll/blog.js
@@ -1,4 +1,4 @@
-import config from "../../js/config.js";
+import config from "../../../js/config.js"
 const url = config.url + "blogpost"
 
 let blogPosts;
